requirements.txt

opencv-python
matplotlib
pytesseract
mysql-connector-python
numpy
Pillow
